Refactored support for additional kinematics
Added Maslow CNC kinematics, a special driver is required for DC-motor control (closed loop). Such a driver is in the pipeline for Tiva C (TM4C123), using DC-motors for XY motion and a stepper for Z. In the pipeline is also a supporting BoosterPack for the Tiva C LaunchPad. Tiva C has been chosen for the two QEI interfaces available.