Added networking plugin
For telnet and/or websocket stream communication. Implemented on top of the lwIP network stack. The ESP32 and MSP432EY401/TM4C129x drivers will soon be updated to use this plugin instead of "local" code.

NOTE: Websocket support is to be regarded as in alpha phase. Simple testing is promising though...