Some optional extensions added
Manual homing of axes with $H command(s), send parser status report on changes (checked for on real-time status request, sent immediately after).

Settings flags for $10 expanded for parser status report, $22 for homing.

The idea for this has been picked up from @shaise's port, but implemented a bit differently in order to avoid breaking existing senders and for flexibility.

Added to core since this is in line with LinuxCNC behaviour.