Serial modified
the serial.c has been modified, mainly the TX function.
Now any data written to the serial is first written to the UART TX FIFO until the FIFO is full, then the TX empty FIFO level interrupt is enabled. When the transmit interrupt is enabled, the data are then written to the TX buffer. When the transmit interrupt is asserted ( TX FIFO almost empty), the TX buffer is written to the TX FIFO until the TX FIFO is full or the TX buffer is empty. the transmit interrupt is deactivated when the TX Buffer is empty.